Apply recommended changes
